Add message handler to node properties
It is now possible to add a on_message callback method to node properties.

By default settable node properties will just update the node data if a
message with a valid payload arrives. Set a on_message callback handler
to process incomming payload and use the HomieNodeProperty.msg_handler as a
template.

The property msg_handler handles payload validation and retained messages.

This will not break existing implemenations that overwrite the
callback() method in node sub-class.